+++ types/email.d.ts
@@ -1,0 +1,8 @@
+export type sendToType = {
+	to: string;
+	subject: string;
+	text: {
+		title: string;
+		msg: string;
+	};
+};
